+++ core/src/test/kotlin/io/github/serpro69/kfaker/FakerServiceTest.kt
@@ -24,7 +24,7 @@ val random = Random()
 internal class FakerServiceTest : DescribeSpec({
     describe("locale for the dictionary") {
         context("it is set to default value") {
-            val dictionary = FakerService(faker = Faker(), random = random).dictionary
+            val dictionary = FakerService(faker = Faker()).dictionary
 
             it("it should contain all keys for 'en' locale") {
                 dictionary.categories.map { it.categoryName.toLowerCase() } shouldContainExactlyInAnyOrder dictionaryKeys
@@ -39,8 +39,8 @@ internal class FakerServiceTest : DescribeSpec({
         }
 
         context("it is set to custom value") {
-            val esDictionary = FakerService(Faker(), Locale.forLanguageTag("es"), random).dictionary
-            val defaultDictionary = FakerService(faker = Faker(), random = random).dictionary
+            val esDictionary = FakerService(Faker(), Locale.forLanguageTag("es")).dictionary
+            val defaultDictionary = FakerService(faker = Faker()).dictionary
 
             it("matching keys should be overwritten in the localized dictionary") {
                 val esAddress = esDictionary.getCategoryByName("address")
@@ -57,7 +57,7 @@ internal class FakerServiceTest : DescribeSpec({
             }
 
             it("partially localized functions with secondary_key should contain non-localized default values ") {
-                val deDict = FakerService(Faker(), Locale.forLanguageTag("de"), random).dictionary
+                val deDict = FakerService(Faker(), Locale.forLanguageTag("de")).dictionary
 
                 val deCommerce = deDict.getCategoryByName("commerce").values
                 val defaultCommerce = defaultDictionary.getCategoryByName("commerce").values
@@ -70,7 +70,7 @@ internal class FakerServiceTest : DescribeSpec({
 
         context("it is set with a valid String value") {
             it("localized dictionary should be loaded") {
-                val esDictionary = FakerService(Faker(), "es", random).dictionary
+                val esDictionary = FakerService(Faker(), "es").dictionary
                 esDictionary shouldNotBe null
             }
         }
@@ -78,7 +78,7 @@ internal class FakerServiceTest : DescribeSpec({
         context("it is set with invalid String value") {
             it("an exception is thrown when loading the dictionary") {
                 val exception = shouldThrow<IllegalArgumentException> {
-                    FakerService(Faker(), "pe", random).dictionary
+                    FakerService(Faker(), "pe").dictionary
                 }
 
                 exception.message shouldBe "Dictionary file not found for locale values: 'pe' or 'pe'"
@@ -86,7 +86,7 @@ internal class FakerServiceTest : DescribeSpec({
         }
 
         context("it is set as `lang-COUNTRY` but dictionary file exists only for `lang`") {
-            val frFRDict = FakerService(Faker(), "fr-FR", random).dictionary
+            val frFRDict = FakerService(Faker(), "fr-FR").dictionary
 
             it("localized dictionary for `lang` should be loaded") {
                 frFRDict shouldNotBe null
@@ -94,7 +94,7 @@ internal class FakerServiceTest : DescribeSpec({
         }
 
         context("it is set as `lang_COUNTRY` String") {
-            val frFRDict = FakerService(Faker(), "fr_FR", random).dictionary
+            val frFRDict = FakerService(Faker(), "fr_FR").dictionary
 
             it("it should be set as `lang-COUNTRY` String") {
                 frFRDict shouldNotBe null
@@ -103,7 +103,7 @@ internal class FakerServiceTest : DescribeSpec({
     }
 
     describe("dictionary is loaded") {
-        val fakerService = FakerService(faker = Faker(), random = random)
+        val fakerService = FakerService(faker = Faker())
 
         context("fetching category by key") {
             val category = fakerService.fetchCategory(CategoryName.ADDRESS)